Kidney-Stone-Detection/requirements.txt
absl-py==1.2.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.2
colorama==0.4.4
cycler==0.11.0
distlib==0.3.4
docopt==0.6.2
filelock==3.6.0
Flask==1.1.1
flatbuffers==2.0.7
fonttools==4.31.2
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
gunicorn==19.9.0
h5py==3.7.0
idna==3.4
importlib-metadata==4.11.3
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==1.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
oauthlib==3.2.1
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pipreqs==0.4.11
platformdirs==2.5.1
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.14.0
Werkzeug==2.2.2
wrapt==1.14.1
xgboost==1.5.2
yarg==0.1.9
zipp==3.8.0
